fix layernorm bug

+++ lib/layernorm.h
@@ -7,6 +7,6 @@
 
 void layernorm_A(Matmul *x, int rows, int cols,
                  const std::vector<__fp16> &gamma,
-                 const std::vector<__fp16> &beta, __fp16 eps = 1e-5);
+                 const std::vector<__fp16> &beta, float eps = 1e-5);
 
 #endif  // _LIB_LAYERNORM_H_
